refactor: Extract toLocatorOptions to its own module
